+++ core/lib/libtomcrypt/sub.mk
@@ -15,3 +15,4 @@ srcs-$(CFG_CRYPTO_CTR) += ctr.c
 srcs-$(CFG_CRYPTO_XTS) += xts.c
 srcs-$(CFG_CRYPTO_CCM) += ccm.c
 srcs-$(CFG_CRYPTO_AES_GCM_FROM_CRYPTOLIB) += gcm.c
+srcs-$(CFG_CRYPTO_DSA) += dsa.c
